Unify `src` module of imagenet and imagenet16.
