--- imagenet/src/utils/model.py
@@ -1,38 +1,0 @@
-import shutil
-
-import torch
-from torchvision import models
-
-
-def save_checkpoint(state, is_best, param_path, epoch):
-    filename = param_path + "checkpoint.pth.tar"
-    torch.save(state, filename)
-    if is_best:
-        shutil.copyfile(filename, param_path + "model_best.pth.tar")
-
-
-def save_model(state, param_path, epoch):
-    filename = param_path + "epoch_{}.pth.tar".format(epoch)
-    torch.save(state, filename)
-
-
-def load_model(model_path, arch="alexnet", paralell=True):
-    """Load model.
-    Args:
-        model_path: Path to the pytorch saved file of the model you want to use
-        arch: Architecture of CNN
-        paralell: False if you want to disable DataParallel
-
-    Returns: CNN model
-    """
-    checkpoint = torch.load(model_path, map_location="cuda:0")
-    model = models.__dict__[arch]()
-    try:
-        model.load_state_dict(checkpoint["state_dict"])
-    except RuntimeError:
-        model.features = torch.nn.DataParallel(model.features)
-        model.load_state_dict(checkpoint["state_dict"])
-        if not paralell:
-            model.features = model.features.module
-            # TODO: This part is different when a model is "resnet".
-    return model
