+++ imagenet16/README.md
@@ -6,7 +6,7 @@ Blur-training with 16-class-ImageNet
 **Default: AlexNet (16 class)**  
 Since the number of class is 16, I change the number of final units from 1000 to 16.
 See more details in `notebooks/models.ipynb`  
-You can also use another architecture by using `--arch [ARCHITECTURE NAME]`. See `python src/train.py -h` for the available models (from pytorchvision's model zoo).
+You can also use another architecture by using `--arch [ARCHITECTURE NAME]`. See `python train.py -h` for the available models (from pytorchvision's model zoo).
 
 
 ## Preparations
@@ -28,7 +28,7 @@ If not, Download the ImageNet dataset from http://www.image-net.org/
 ## run examples
 General usage example:
 ```bash
-$ python src/train.py --arch [ARCHITECTURE NAME] --mode [TRAINING MODE] -n [EXPERIMENT NAME] 
+$ python train.py --arch [ARCHITECTURE NAME] --mode [TRAINING MODE] -n [EXPERIMENT NAME] 
 ```  
 
 For `train.py`, you need to use `--exp_name` or `-n` option to define your experiment's name.
@@ -42,49 +42,49 @@ by using `--mode [TRAINING MODE]` option.
 This mode trains Normal model (default: AlexNet).  
 usage example:  
 ```bash
-$ python src/train.py --mode normal -e 60 -b 64 --lr 0.01 -n normal
+$ python train.py --mode normal -e 60 -b 64 --lr 0.01 -n normal
 ```
 
 - **all**  
 This mode blurs ALL images in the training mode.  
 usage example:  
 ```bash
-$ python src/train.py --mode all -s1 -n all_s1
+$ python train.py --mode all -s1 -n all_s1
 ```
 
 - **mix**    
 This mode blurs half training data.
 usage example:  
 ```bash
-$ python src/train.py --mode mix -s 1 -n mix_s1
+$ python train.py --mode mix -s 1 -n mix_s1
 ```
 
 - **random-mix** <br>
 This mode blurs half training data **randomly**. <br>
 usage example:
 ```bash
-$ python src/train.py --arch alexnet --mode random-mix --min_sigma 0 --max_sigma 5 -n alexnet_random-mix_s0-5
+$ python train.py --arch alexnet --mode random-mix --min_sigma 0 --max_sigma 5 -n alexnet_random-mix_s0-5
 ```
 
 - **single-step**    
 This mode blurs first half epochs (e.g. first 30 epochs in 60 entire epochs) in the training.
 usage example:  
 ```bash
-$ python src/train.py --mode single-step -s 1 -n single-step_s1
+$ python train.py --mode single-step -s 1 -n single-step_s1
 ```
 
 - **multi-steps**  
 This mode blurs images step by step (e.g. every 10 epochs).  
 usage example:  
 ```bash
-$ python src/train.py --mode multi-steps -n multi-steps
+$ python train.py --mode multi-steps -n multi-steps
 ```
 
 - `--resume [PATH TO SAVED MODEL]`   
 This option trains your saved model starting from the latest epoch.  
 usage example:  
 ```bash
-$ python src/train.py --arch alexnet --mode mix -s 1 -n alexnet_mix_s1 --resume ../logs/models/alexnet_mix_s1/checkpoint.pth.tar 
+$ python train.py --arch alexnet --mode mix -s 1 -n alexnet_mix_s1 --resume ../logs/models/alexnet_mix_s1/checkpoint.pth.tar 
 ```
