+++ src/analysis/lowpass/plot.py
