+++ src/analysis/rsa/bandpass/mean_rdms.py
@@ -10,7 +10,7 @@ from scipy.spatial.distance import squareform, pdist
 
 # add the path to load src module
 current_dir = pathlib.Path(os.path.abspath(__file__)).parent
-sys.path.append(os.path.join(str(current_dir), "../../../../"))
+sys.path.append(os.path.join(str(current_dir), "../../../../imagenet16/"))
 
 from src.analysis.rsa.rdm import alexnet_layers
 from src.analysis.rsa.activations import load_activations
