+++ src/analysis/rsa/bandpass/mlr.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 from sklearn import linear_model
 
-sys.path.append("../../../../")
+sys.path.append("../../../../imagenet16/")
 
 from src.analysis.rsa.rdm import alexnet_layers
 from src.analysis.rsa.activations import load_activations
