+++ src/analysis/rsa/plot_mean_rdms.py
@@ -8,7 +8,7 @@ from matplotlib import pyplot as plt
 
 # add the path to load src module
 current_dir = pathlib.Path(os.path.abspath(__file__)).parent
-sys.path.append(os.path.join(str(current_dir), "../../../"))
+sys.path.append(os.path.join(str(current_dir), "../../../imagenet16/"))
 
 from src.analysis.rsa.rdm import alexnet_layers
