+++ src/analysis/shape_bias/compute_shape_bias.py
