+++ src/analysis/shape_bias/plot_shape-bias_random-mix.py
