+++ src/dataset/info/wordnet.is_a.txt
