+++ src/image_process/bandpass_images.py
@@ -6,7 +6,7 @@ import torch
 
 # add a path to load src module
 current_dir = pathlib.Path(os.path.abspath(__file__)).parent
-sys.path.append(os.path.join(str(current_dir), "../../"))
+sys.path.append(os.path.join(str(current_dir), "../../imagenet16/"))
 
 from src.dataset.imagenet16 import (
     make_test_images_by_class,
